Move run_coverage
